upload multiple images & cleaning unnecessery gallery form - clear version
