+++ README.md
@@ -56,8 +56,6 @@
     
     - GALERIA:
 
-        -przesyłanie kilku zdjęć naraz
-
         -osobny plik na widgety
 
         -wybieranie main image w galerii (przy pomocy radioset)
