+++ templates/blog/post_detail.html
@@ -16,12 +16,6 @@
   {% if photos %}
     <h3>{% trans "Gallery" %}</h3>
 
-<!--    {% if gallery.gallery_title  %}-->
-<!--        <h2>{{ gallery.gallery_title }}</h2>-->
-<!--    {% else %}-->
-<!--        <p>No gallery title available</p>-->
-<!--    {% endif %}-->
-
     <div class="post-gallery">
       {% for photo in photos %}
         <div class="photo">
